+++ bin/dev-tools.php
@@ -20,6 +20,7 @@ declare(strict_types=1);
 namespace FastForward\DevTools;
 
 use FastForward\DevTools\Console\DevTools;
+use FastForward\DevTools\Console\DevToolsComposer;
 use Symfony\Component\Console\Input\ArgvInput;
 
 $projectVendorAutoload = \dirname(__DIR__, 4) . '/vendor/autoload.php';
@@ -27,4 +28,11 @@ $pluginVendorAutoload = \dirname(__DIR__) . '/vendor/autoload.php';
 
 require_once file_exists($projectVendorAutoload) ? $projectVendorAutoload : $pluginVendorAutoload;
 
-DevTools::create()->run(new ArgvInput([...$argv, '--no-plugins']));
+$input = new ArgvInput([...$argv, '--no-plugins']);
+
+$command = $input->getFirstArgument();
+$application = (null !== $command && DevTools::create()->has($command))
+    ? DevTools::create()
+    : DevToolsComposer::create();
+
+$application->run($input);