+++ src/Console/Command/AgentsCommand.php
@@ -20,13 +20,13 @@ declare(strict_types=1);
 namespace FastForward\DevTools\Console\Command;
 
 use FastForward\DevTools\Console\Command\Traits\LogsCommandResults;
-use Composer\Command\BaseCommand;
 use FastForward\DevTools\Console\Input\HasJsonOption;
 use FastForward\DevTools\Filesystem\FilesystemInterface;
 use FastForward\DevTools\Path\DevToolsPathResolver;
 use FastForward\DevTools\Sync\PackagedDirectorySynchronizer;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -34,7 +34,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  * Synchronizes packaged Fast Forward project agents into the consumer repository.
  */
 #[AsCommand(name: 'agents', description: 'Synchronizes Fast Forward project agents into .agents/agents directory.')]
-final class AgentsCommand extends BaseCommand implements LoggerAwareCommandInterface
+final class AgentsCommand extends Command implements LoggerAwareCommandInterface
 {
     use HasJsonOption;
     use LogsCommandResults;
@@ -51,7 +51,7 @@ final class AgentsCommand extends BaseCommand implements LoggerAwareCommandInter
         private readonly FilesystemInterface $filesystem,
         private readonly LoggerInterface $logger,
     ) {
-        parent::__construct();
+        parent::__construct('agents');
     }
 
     /**
@@ -98,7 +98,7 @@ final class AgentsCommand extends BaseCommand implements LoggerAwareCommandInter
             $this->logger->info('Created .agents/agents directory.');
         }
 
-        $this->synchronizer->setLogger($this->getIO());
+        $this->synchronizer->setLogger($this->logger);
 
         $result = $this->synchronizer->synchronize($agentsDir, $packageAgentsPath, self::AGENTS_DIRECTORY);