+++ src/ServiceProvider/DevToolsServiceProvider.php
@@ -34,7 +34,9 @@ use FastForward\DevTools\Changelog\Renderer\MarkdownRendererInterface;
 use FastForward\DevTools\Changelog\Checker\UnreleasedEntryChecker;
 use FastForward\DevTools\Changelog\Checker\UnreleasedEntryCheckerInterface;
 use FastForward\DevTools\Console\CommandLoader\DevToolsCommandLoader;
+use FastForward\DevTools\Console\CommandLoader\SymfonyDevToolsCommandLoader;
 use FastForward\DevTools\Console\Formatter\LogLevelOutputFormatter;
+use FastForward\DevTools\Console\DevTools;
 use FastForward\DevTools\Console\Logger\OutputFormatLogger;
 use FastForward\DevTools\Console\Logger\Processor\CommandInputProcessor;
 use FastForward\DevTools\Console\Logger\Processor\CommandOutputProcessor;
@@ -84,6 +86,7 @@ use FastForward\DevTools\Process\ProcessQueue;
 use FastForward\DevTools\Process\ProcessQueueInterface;
 use FastForward\DevTools\Process\XdebugDisablingProcessEnvironmentConfigurator;
 use FastForward\DevTools\Path\DevToolsPathResolver;
+use FastForward\DevTools\Console\DevToolsComposer;
 use FastForward\DevTools\Path\WorkingProjectPathResolver;
 use FastForward\DevTools\Psr\Clock\SystemClock;
 use FastForward\DevTools\Resource\DifferInterface;
@@ -158,6 +161,8 @@ final class DevToolsServiceProvider implements ServiceProviderInterface
 
             // Console
             CommandLoaderInterface::class => get(DevToolsCommandLoader::class),
+            DevTools::class => create(DevTools::class)->constructor(get(SymfonyDevToolsCommandLoader::class)),
+            DevToolsComposer::class => create(DevToolsComposer::class)->constructor(get(DevToolsCommandLoader::class)),
             CommandProvider::class => get(DevToolsCommandProvider::class),
             ConsoleOutputInterface::class => create(ConsoleOutput::class)
                 ->method('setVerbosity', ConsoleOutputInterface::VERBOSITY_VERBOSE)